+++ data-raw/cds_datasets.R
@@ -5,13 +5,13 @@ library(DataSpaceR)
 con <- connectDS()
 
 cvd812 <- con$getStudy('cvd812')
-CAVD812_mAB <- cvd812$getDataset("NAB Ig") %>%
+CAVD812_mAB <- cvd812$getDataset("NAB Ig") |>
   distinct(
     product = mab_name_source,
     virus = cds_virus_name,
     ic50 = titer_curve_ic50,
     ic80 = titer_curve_ic80
-  ) %>% as.data.frame()
+  ) |> as.data.frame()
 
 usethis::use_data(CAVD812_mAB, overwrite = TRUE)